Add email code logic and move two_factor into separate modules
