+++ src/db/models/two_factor.rs
@@ -1,4 +1,11 @@
+use diesel;
+use diesel::prelude::*;
 use serde_json::Value;
+
+use crate::api::EmptyResult;
+use crate::db::DbConn;
+use crate::db::schema::twofactor;
+use crate::error::MapResult;
 
 use super::User;
 
@@ -28,6 +35,8 @@ pub enum TwoFactorType {
     // These are implementation details
     U2fRegisterChallenge = 1000,
     U2fLoginChallenge = 1001,
+    EmailVerificationChallenge = 1002,
+
 }
 
 /// Local methods
@@ -58,14 +67,6 @@ impl TwoFactor {
         })
     }
 }
-
-use crate::db::schema::twofactor;
-use crate::db::DbConn;
-use diesel;
-use diesel::prelude::*;
-
-use crate::api::EmptyResult;
-use crate::error::MapResult;
 
 /// Database methods
 impl TwoFactor {
